backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
torch>=2.2.0
torchvision>=0.17.0
ftfy>=6.1.1
regex>=2023.10.3
tqdm>=4.66.1
git+https://github.com/openai/CLIP.git
qdrant-client>=1.7.0
werkzeug==3.0.1
numpy>=1.24.3
